// src/interfaces/AuthInterface.tsx
export interface loginData {
  username: string;
  password: string;
}

// Generated by https://quicktype.io

export interface Login {
  token_type?:    string|null|undefined;
  expires_in?:    number|null|undefined;
  access_token?:  string|null|undefined;
  refresh_token?: string|null|undefined;
  data?:     string|null|undefined;
  messages?: string|string[]|null|undefined;
}

// Generated by https://quicktype.io

export interface Session {
  usuario: Usuario;
}

export interface Usuario {
  id:                     number;
  nombre:                 string;
  email:                  string;
  identificacion_usuario: string;
  telefono:               string;
  asociado:               Asociado;
  rol:                    Rol;
  permisos:               UsuarioPermiso[];
}

export interface Asociado {
  id:               number;
  nombre:           string;
  numero_documento: string;
}

export interface UsuarioPermiso {
  id:         number;
  nombre:     string;
  posicion:   number;
  icono_menu: string;
  type:       string;
  opciones:   Opciones[];
}

export interface Opciones {
  id:         number;
  nombre:     string;
  posicion:   number;
  icono_menu: string|null;
  url:        string;
  url_ayuda:  null;
  type:       string;
  permisos:   OpcionesPermisos[];
}

export interface OpcionesPermisos {
  nombre:    string;
  titulo:    string;
  permitido: boolean;
  id:        number;
}

export interface Rol {
  id:     number;
  nombre: string;
  tipo:   string;
}